:sparkles: Refactor package
